requirements.txt
altair==5.5.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
Flask==3.1.1
Flask-Mail==0.10.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.27.1
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pillow==10.2.0
protobuf==4.25.6
pyarrow==19.0.1
pycparser==2.22
pydeck==0.9.1
pygame==2.5.2
Pygments==2.19.1
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
pyzbar==0.1.9
qrcode==7.4.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
setuptools==76.0.0
six==1.17.0
smmap==5.0.2
streamlit==1.44.1
tenacity==8.5.0
toml==0.10.2
tornado==6.4.2
ttkthemes==3.2.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.45.1
